More fixes for non-standard users
Needed to fix some python code that was added that works only on Python 3.
The top level directories for the registry should be owned by
root and be world readable/executable so that different users
can read the registry.

cherry-pick from 79e8a8e99a3c9c6d16b63bb5d4b2b18a4b828b11

+++ base/server/python/pki/server/deployment/scriptlets/instance_layout.py
@@ -214,8 +214,9 @@ class PkiScriptlet(pkiscriptlet.AbstractBasePkiScriptlet):
             deployer.mdict['pki_instance_registry_path'])
         # remove Tomcat PKI registry (if empty)
         if deployer.instance.tomcat_instances() == 0:
-            deployer.directory.delete(
-                deployer.mdict['pki_instance_type_registry_path'])
+            if deployer.mdict['pki_root_prefix'] != "":
+                deployer.directory.delete(
+                    deployer.mdict['pki_instance_type_registry_path'])
 
 
 # Callback only when the /usr/share/pki/server/conf directory
